backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
redis
rq
apscheduler
python-multipart
requests
dnspython
